[build] Convert to a proper build script
